Split end-to-end test classes
