+++ test/Cocona.Test/Integration/CoconaAppRunTest.cs
@@ -8,7 +8,6 @@ using System.Threading.Tasks;
 using Cocona.Application;
 using Cocona.Command.BuiltIn;
 using Cocona.CommandLine;
-using Cocona.Filters;
 using Cocona.ShellCompletion.Candidate;
 using FluentAssertions;
 using Xunit;
@@ -19,181 +18,8 @@ using CoconaApp = Cocona.CoconaLiteApp;
 
 namespace Cocona.Test.Integration
 {
-    [Collection("End to End")] // NOTE: Test cases use `Console` and does not run in parallel.
-    public class EndToEndTest
+    public class CoconaAppRunTest : EndToEndTestBase
     {
-        public enum RunBuilderMode
-        {
-            CreateBuilder,
-            CreateHostBuilder,
-            Shortcut,
-        }
-        private (string StandardOut, string StandardError, int ExitCode) Run(string[] args, Action<string[]> action)
-        {
-            var stdOutWriter = new StringWriter();
-            var stdErrWriter = new StringWriter();
-
-            Console.SetOut(stdOutWriter);
-            Console.SetError(stdErrWriter);
-
-            action(args);
-
-            return (stdOutWriter.ToString(), stdErrWriter.ToString(), Environment.ExitCode);
-        }
-
-        private async Task<(string StandardOut, string StandardError, int ExitCode)> RunAsync(string[] args, Func<string[], Task> action)
-        {
-            var stdOutWriter = new StringWriter();
-            var stdErrWriter = new StringWriter();
-
-            Console.SetOut(stdOutWriter);
-            Console.SetError(stdErrWriter);
-
-            await action(args);
-
-            return (stdOutWriter.ToString(), stdErrWriter.ToString(), Environment.ExitCode);
-        }
-
-        private (string StandardOut, string StandardError, int ExitCode) Run<T>(RunBuilderMode mode, string[] args)
-        {
-            var stdOutWriter = new StringWriter();
-            var stdErrWriter = new StringWriter();
-
-            Console.SetOut(stdOutWriter);
-            Console.SetError(stdErrWriter);
-
-            switch (mode)
-            {
-                case RunBuilderMode.CreateBuilder:
-                    var builder = CoconaApp.CreateBuilder(args);
-                    var app = builder.Build();
-                    app.AddCommands<T>();
-
-                    app.Run();
-                    break;
-                case RunBuilderMode.CreateHostBuilder:
-                    CoconaApp.CreateHostBuilder()
-                        .Run<T>(args);
-                    break;
-                case RunBuilderMode.Shortcut:
-                    CoconaApp.Run<T>(args);
-                    break;
-            }
-
-            return (stdOutWriter.ToString(), stdErrWriter.ToString(), Environment.ExitCode);
-        }
-
-        private async Task<(string StandardOut, string StandardError, int ExitCode)> RunAsync<T>(RunBuilderMode mode, string[] args, CancellationToken cancellationToken)
-        {
-            var stdOutWriter = new StringWriter();
-            var stdErrWriter = new StringWriter();
-
-            Console.SetOut(stdOutWriter);
-            Console.SetError(stdErrWriter);
-
-            switch (mode)
-            {
-                case RunBuilderMode.CreateBuilder:
-                    var builder = CoconaApp.CreateBuilder(args);
-                    var app = builder.Build();
-                    app.AddCommands<T>();
-
-                    await app.RunAsync(cancellationToken);
-                    break;
-                case RunBuilderMode.CreateHostBuilder:
-                    await CoconaApp.CreateHostBuilder()
-                        .RunAsync<T>(args, cancellationToken: cancellationToken);
-                    break;
-                case RunBuilderMode.Shortcut:
-                    await CoconaApp.RunAsync<T>(args, cancellationToken: cancellationToken);
-                    break;
-            }
-
-            return (stdOutWriter.ToString(), stdErrWriter.ToString(), Environment.ExitCode);
-        }
-
-        private (string StandardOut, string StandardError, int ExitCode) Run(RunBuilderMode mode, string[] args, Type[] types)
-        {
-            var stdOutWriter = new StringWriter();
-            var stdErrWriter = new StringWriter();
-
-            Console.SetOut(stdOutWriter);
-            Console.SetError(stdErrWriter);
-
-            switch (mode)
-            {
-                case RunBuilderMode.CreateBuilder:
-                    var builder = CoconaApp.CreateBuilder(args);
-                    var app = builder.Build();
-                    foreach (var type in types)
-                    {
-                        app.AddCommands(type);
-                    }
-
-                    app.Run();
-                    break;
-                case RunBuilderMode.CreateHostBuilder:
-                    CoconaApp.CreateHostBuilder()
-                        .Run(args, types);
-                    break;
-                case RunBuilderMode.Shortcut:
-                    CoconaApp.Run(args, types);
-                    break;
-            }
-
-            return (stdOutWriter.ToString(), stdErrWriter.ToString(), Environment.ExitCode);
-        }
-
-        private async Task<(string StandardOut, string StandardError, int ExitCode)> RunAsync<T>(RunBuilderMode mode, string[] args, Type[] types, CancellationToken cancellationToken)
-        {
-            var stdOutWriter = new StringWriter();
-            var stdErrWriter = new StringWriter();
-
-            Console.SetOut(stdOutWriter);
-            Console.SetError(stdErrWriter);
-
-            switch (mode)
-            {
-                case RunBuilderMode.CreateBuilder:
-                    var builder = CoconaApp.CreateBuilder(args);
-                    var app = builder.Build();
-                    foreach (var type in types)
-                    {
-                        app.AddCommands(type);
-                    }
-
-                    await app.RunAsync(cancellationToken);
-                    break;
-                case RunBuilderMode.CreateHostBuilder:
-                    await CoconaApp.CreateHostBuilder()
-                        .RunAsync(args, types, cancellationToken: cancellationToken);
-                    break;
-                case RunBuilderMode.Shortcut:
-                    await CoconaApp.RunAsync(args, types, cancellationToken: cancellationToken);
-                    break;
-            }
-
-            return (stdOutWriter.ToString(), stdErrWriter.ToString(), Environment.ExitCode);
-        }
-
-        private (string StandardOut, string StandardError, int ExitCode) Run(string[] args, Delegate[] delegates)
-        {
-            var stdOutWriter = new StringWriter();
-            var stdErrWriter = new StringWriter();
-
-            Console.SetOut(stdOutWriter);
-            Console.SetError(stdErrWriter);
-
-            var app = CoconaApp.Create(args);
-            foreach (var @delegate in delegates)
-            {
-                app.AddCommand(@delegate);
-            }
-            app.Run();
-
-            return (stdOutWriter.ToString(), stdErrWriter.ToString(), Environment.ExitCode);
-        }
-
         [Theory]
         [InlineData(RunBuilderMode.CreateHostBuilder)]
         [InlineData(RunBuilderMode.CreateBuilder)]
@@ -932,272 +758,6 @@ namespace Cocona.Test.Integration
         {
             public void A([Option]int a, [Option(StopParsingOptions = true)]string b, [Argument]string arg0, [Argument]string[] args)
                 => Console.WriteLine($"A:{a}:{b}:{arg0}:{string.Join(",", args)}");
-        }
-
-        [Fact]
-        public void ParameterInjection_CoconaApp_CreateBuilder_Delegate()
-        {
-            var (stdOut, stdErr, exitCode) = Run(new[] { "--age", "18" }, args =>
-            {
-                var builder = CoconaApp.CreateBuilder(args);
-#if COCONA_LITE
-                builder.Services.Add(new Lite.ServiceDescriptor(typeof(IMyService), (_, _) => new MyService(), singleton: true));
-#else
-                builder.Services.Add(new Microsoft.Extensions.DependencyInjection.ServiceDescriptor(typeof(IMyService), new MyService()));
-#endif
-                var app = builder.Build();
-                app.AddCommand((int age, IMyService myService) => Console.WriteLine($"Hello {myService.GetName()} ({age})!"));
-                app.Run();
-            });
-
-            stdOut.Should().Be("Hello Alice (18)!" + Environment.NewLine);
-            exitCode.Should().Be(0);
-        }
-
-        [Fact]
-        public void ParameterInjection_CoconaApp_CreateBuilder_Type()
-        {
-            {
-                var (stdOut, stdErr, exitCode) = Run(new[] { "hello-without-from-service", "--age", "18" }, args =>
-                {
-                    var builder = CoconaApp.CreateBuilder(args);
-#if COCONA_LITE
-                    builder.Services.Add(new Lite.ServiceDescriptor(typeof(IMyService), (_, _) => new MyService(), singleton: true));
-#else
-                    builder.Services.Add(new Microsoft.Extensions.DependencyInjection.ServiceDescriptor(typeof(IMyService), new MyService()));
-#endif
-                    var app = builder.Build();
-                    app.AddCommands<ParameterInjectionTestCommands>();
-                    app.Run();
-                });
-
-                stdErr.Should().Contain("'--my-service' is required");
-                exitCode.Should().Be(1);
-            }
-
-            {
-                var (stdOut, stdErr, exitCode) = Run(new[] { "hello-with-from-service", "--age", "18" }, args =>
-                {
-                    var builder = CoconaApp.CreateBuilder(args);
-#if COCONA_LITE
-                    builder.Services.Add(new Lite.ServiceDescriptor(typeof(IMyService), (_, _) => new MyService(), singleton: true));
-#else
-                    builder.Services.Add(new Microsoft.Extensions.DependencyInjection.ServiceDescriptor(typeof(IMyService), new MyService()));
-#endif
-                    var app = builder.Build();
-                    app.AddCommands<ParameterInjectionTestCommands>();
-                    app.Run();
-                });
-
-                stdOut.Should().Be("Hello Alice (18)!" + Environment.NewLine);
-                exitCode.Should().Be(0);
-            }
-        }
-        class ParameterInjectionTestCommands
-        {
-            public void HelloWithoutFromService(int age, IMyService myService) => Console.WriteLine($"Hello {myService.GetName()} ({age})!");
-            public void HelloWithFromService(int age, [FromService]IMyService myService) => Console.WriteLine($"Hello {myService.GetName()} ({age})!");
-        }
-
-        interface IMyService
-        {
-            string GetName();
-        }
-        class MyService : IMyService
-        {
-            public string GetName() => "Alice";
-        }
-
-        [Fact]
-        public void CommandFilter_CommandType_Attributes()
-        {
-            var (stdOut, stdErr, exitCode) = Run(Array.Empty<string>(), args =>
-            {
-                var app = CoconaApp.Create(args);
-                app.AddCommands<TestCommand_HasFilter>();
-                app.Run();
-            });
-
-            stdOut.Should().Be(String.Join(Environment.NewLine, new[] { "Class#Begin", "Method#Begin", "Hello", "Method#End", "Class#End" }) + Environment.NewLine);
-            exitCode.Should().Be(0);
-        }
-
-        [Fact]
-        public void CommandFilter_CommandDelegate()
-        {
-            var (stdOut, stdErr, exitCode) = Run(Array.Empty<string>(), args =>
-            {
-                var app = CoconaApp.Create(args);
-                app.AddCommand([TestCommand_HasFilter.MyCommandFilter("Method")]() => Console.WriteLine("Hello"))
-                    .WithFilter(new TestCommand_HasFilter.MyCommandFilter("Builder"));
-                app.Run();
-            });
-
-            stdOut.Should().Be(String.Join(Environment.NewLine, new[] { "Method#Begin", "Builder#Begin", "Hello", "Builder#End", "Method#End" }) + Environment.NewLine);
-            exitCode.Should().Be(0);
-        }
-
-        [Fact]
-        public void CommandFilter_CommandDelegate_Multiple()
-        {
-            var (stdOut, stdErr, exitCode) = Run(Array.Empty<string>(), args =>
-            {
-                // Filters are applied from the outer side
-                // Class -> Builder2 -> Builder1 -> Method
-                var app = CoconaApp.Create(args);
-                app.AddCommand([TestCommand_HasFilter.MyCommandFilter("Method")]() => Console.WriteLine("Hello"))
-                    .WithFilter(new TestCommand_HasFilter.MyCommandFilter("Builder1"))
-                    .WithFilter(new TestCommand_HasFilter.MyCommandFilter("Builder2"));
-                app.Run();
-            });
-
-            stdOut.Should().Be(String.Join(Environment.NewLine, new[] { "Method#Begin", "Builder1#Begin", "Builder2#Begin", "Hello", "Builder2#End", "Builder1#End", "Method#End" }) + Environment.NewLine);
-            exitCode.Should().Be(0);
-        }
-
-        [Fact]
-        public void CommandFilter_UseFilter_CommandDelegate()
-        {
-            var (stdOut, stdErr, exitCode) = Run(Array.Empty<string>(), args =>
-            {
-                // Filters are applied from the outer side
-                // Builder0 (UseFilter) -> Method -> Builder1 -> Builder2
-                var app = CoconaApp.Create(args);
-                app.UseFilter(new TestCommand_HasFilter.MyCommandFilter("Builder0"));
-                app.AddCommand([TestCommand_HasFilter.MyCommandFilter("Method")]() => Console.WriteLine("Hello")) // Same as .WithFilter()
-                    .WithFilter(new TestCommand_HasFilter.MyCommandFilter("Builder1"))
-                    .WithFilter(new TestCommand_HasFilter.MyCommandFilter("Builder2")); // Method, Builder1, Builder2
-                app.Run();
-            });
-
-            stdOut.Should().Be(String.Join(Environment.NewLine, new[] { "Builder0#Begin", "Method#Begin", "Builder1#Begin", "Builder2#Begin", "Hello", "Builder2#End", "Builder1#End", "Method#End", "Builder0#End" }) + Environment.NewLine);
-            exitCode.Should().Be(0);
-        }
-
-        [Fact]
-        public void CommandFilter_UseFilter_CommandType()
-        {
-            var (stdOut, stdErr, exitCode) = Run(Array.Empty<string>(), args =>
-            {
-                // Filters are applied from the outer side
-                // Builder0 (UseFilter) -> Class -> Method
-                var app = CoconaApp.Create(args);
-                app.UseFilter(new TestCommand_HasFilter.MyCommandFilter("Builder0"));
-                app.UseFilter(new TestCommand_HasFilter.MyCommandFilter("Builder1"));
-                app.AddCommands<TestCommand_HasFilter>();
-                app.Run();
-            });
-
-            stdOut.Should().Be(String.Join(Environment.NewLine, new[] { "Builder0#Begin", "Builder1#Begin", "Class#Begin", "Method#Begin", "Hello", "Method#End", "Class#End", "Builder1#End", "Builder0#End" }) + Environment.NewLine);
-            exitCode.Should().Be(0);
-        }
-
-        [Fact]
-        public void CommandFilter_UseFilter_CommandDelegate_SubCommand()
-        {
-            var (stdOut, stdErr, exitCode) = Run(new [] { "sub-command", "command1" }, args =>
-            {
-                // Filters are applied from the outer side
-                // Builder0 (UseFilter) -> Class -> Method
-                var app = CoconaApp.Create(args);
-                app.UseFilter(new TestCommand_HasFilter.MyCommandFilter("Builder0"));
-                app.UseFilter(new TestCommand_HasFilter.MyCommandFilter("Builder1"));
-                app.AddSubCommand("sub-command", x =>
-                {
-                    x.UseFilter(new TestCommand_HasFilter.MyCommandFilter("Builder2"));
-                    x.AddCommand("command1", () => Console.WriteLine("Hello"));
-                    x.AddCommand("command2", () => Console.WriteLine("A"));
-                });
-
-                app.UseFilter(new TestCommand_HasFilter.MyCommandFilter("Builder3"));
-                app.AddSubCommand("sub-command-2", x =>
-                {
-                    app.UseFilter(new TestCommand_HasFilter.MyCommandFilter("Builder4"));
-                    x.AddCommand("command1", () => Console.WriteLine("B"));
-                    x.AddCommand("command2", () => Console.WriteLine("C"));
-                });
-                app.Run();
-            });
-
-            stdOut.Should().Be(String.Join(Environment.NewLine, new[] { "Builder0#Begin", "Builder1#Begin", "Builder2#Begin", "Hello", "Builder2#End", "Builder1#End", "Builder0#End" }) + Environment.NewLine);
-            exitCode.Should().Be(0);
-        }
-
-        [Fact]
-        public void CommandFilter_UseFilter_CommandType_SubCommand()
-        {
-            var (stdOut, stdErr, exitCode) = Run(new [] { "sub-command", "hello" }, args =>
-            {
-                // Filters are applied from the outer side
-                // Builder0 (UseFilter) -> Class -> Method
-                var app = CoconaApp.Create(args);
-                app.UseFilter(new TestCommand_HasFilter.MyCommandFilter("Builder0"));
-                app.UseFilter(new TestCommand_HasFilter.MyCommandFilter("Builder1"));
-                app.AddSubCommand("sub-command", x =>
-                {
-                    x.UseFilter(new TestCommand_HasFilter.MyCommandFilter("Builder2"));
-                    x.AddCommands<TestCommandMany_HasFilter>();
-                });
-
-                app.UseFilter(new TestCommand_HasFilter.MyCommandFilter("Builder3"));
-                app.AddSubCommand("sub-command-2", x =>
-                {
-                    app.UseFilter(new TestCommand_HasFilter.MyCommandFilter("Builder4"));
-                    x.AddCommand("command1", () => Console.WriteLine("A"));
-                    x.AddCommand("command2", () => Console.WriteLine("B"));
-                });
-                app.Run();
-            });
-
-            stdOut.Should().Be(String.Join(Environment.NewLine, new[] { "Builder0#Begin", "Builder1#Begin", "Builder2#Begin", "Class#Begin", "Method#Begin", "Hello", "Method#End", "Class#End", "Builder2#End", "Builder1#End", "Builder0#End" }) + Environment.NewLine);
-            exitCode.Should().Be(0);
-        }
-
-        [MyCommandFilter("Class")]
-        class TestCommand_HasFilter
-        {
-            [MyCommandFilter("Method")]
-            public void Hello() => Console.WriteLine("Hello");
-
-            public class MyCommandFilter : CommandFilterAttribute
-            {
-                private string _label;
-                public MyCommandFilter(string label)
-                {
-                    _label = label;
-                }
-                public override async ValueTask<int> OnCommandExecutionAsync(CoconaCommandExecutingContext ctx, CommandExecutionDelegate next)
-                {
-                    Console.WriteLine($"{_label}#Begin");
-                    var result = await next(ctx);
-                    Console.WriteLine($"{_label}#End");
-                    return result;
-                }
-            }
-        }
-
-        [MyCommandFilter("Class")]
-        class TestCommandMany_HasFilter
-        {
-            [MyCommandFilter("Method")]
-            public void Hello() => Console.WriteLine("Hello");
-            public void Konnichiwa() => Console.WriteLine("Konnichiwa");
-
-            public class MyCommandFilter : CommandFilterAttribute
-            {
-                private string _label;
-                public MyCommandFilter(string label)
-                {
-                    _label = label;
-                }
-                public override async ValueTask<int> OnCommandExecutionAsync(CoconaCommandExecutingContext ctx, CommandExecutionDelegate next)
-                {
-                    Console.WriteLine($"{_label}#Begin");
-                    var result = await next(ctx);
-                    Console.WriteLine($"{_label}#End");
-                    return result;
-                }
-            }
         }
     }
 }
